# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Application Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ConnectEmailBody {
  /** Email provider slug: gmail, outlook, yahoo, custom */
  provider: string;
  email: string;
  /**
   * App password or IMAP password (required unless oauthToken is provided)
   * @nullable
   */
  password?: string | null;
  /**
   * Google OAuth access token (used instead of password for Gmail OAuth sign-in)
   * @nullable
   */
  oauthToken?: string | null;
  /**
   * Required for custom provider
   * @nullable
   */
  imapHost?: string | null;
  /**
   * Required for custom provider (e.g. 993)
   * @nullable
   */
  imapPort?: number | null;
}

export interface EmailAccountItem {
  id: number;
  email: string;
  provider: string;
  /**
   * ISO 8601 timestamp of last scan
   * @nullable
   */
  lastScanned: string | null;
}

export interface EmailStatusResponse {
  accounts: EmailAccountItem[];
}

export interface DisconnectEmailBody {
  /**
   * ID of the specific session to disconnect. If omitted, all accounts are disconnected.
   * @nullable
   */
  sessionId?: number | null;
}

export interface DisconnectResponse {
  success: boolean;
}

export interface ScanEmailsBody {
  /**
   * Maximum number of emails to scan (default 200)
   * @nullable
   */
  maxEmails?: number | null;
  /**
   * How many days back to look (default 90)
   * @nullable
   */
  daysBack?: number | null;
  /**
   * If true, delete applications outside the scan window before inserting new results (default true)
   * @nullable
   */
  clearPrevious?: boolean | null;
  /**
   * Fresh Google OAuth access token (required when session uses OAuth auth)
   * @nullable
   */
  oauthToken?: string | null;
}

export interface ScanEmailsResponse {
  /** Number of job-related emails found */
  found: number;
  /** Number of new applications added */
  added: number;
  /** Number of existing applications updated */
  updated: number;
}

export interface Application {
  id: number;
  /** ISO 8601 date string */
  dateOfContact: string;
  /** @nullable */
  position: string | null;
  /** @nullable */
  employer: string | null;
  /**
   * Name of recruiter or hiring manager
   * @nullable
   */
  contactName: string | null;
  /**
   * For interview emails: interviewer name(s) and title(s)
   * @nullable
   */
  interviewerInfo?: string | null;
  /** email, zoom, teams, google-meet, phone, linkedin, other */
  methodOfContact: string;
  /** @nullable */
  emailAddress: string | null;
  /** interview, next-stage, rejected, no-response */
  result: string;
  /** @nullable */
  notes: string | null;
  /**
   * Original email message ID for deduplication
   * @nullable
   */
  sourceEmailId: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateApplicationBody {
  dateOfContact: string;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  employer?: string | null;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  interviewerInfo?: string | null;
  methodOfContact: string;
  /** @nullable */
  emailAddress?: string | null;
  result: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateApplicationBody {
  dateOfContact?: string;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  employer?: string | null;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  interviewerInfo?: string | null;
  methodOfContact?: string;
  /** @nullable */
  emailAddress?: string | null;
  result?: string;
  /** @nullable */
  notes?: string | null;
}

export type ApplicationsSummaryByResult = {
  interview: number;
  nextStage: number;
  rejected: number;
  noResponse: number;
};

export interface ApplicationsSummary {
  total: number;
  byResult: ApplicationsSummaryByResult;
  recentActivity: Application[];
}

export type ListApplicationsParams = {
  result?: string;
  search?: string;
};