# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Job Application Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: email
    description: Email provider connection and scanning
  - name: applications
    description: Job application CRUD operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /email/connect:
    post:
      operationId: connectEmail
      tags: [email]
      summary: Connect to email provider
      description: Connect to an IMAP email provider using credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectEmailBody"
      responses:
        "200":
          description: Connected successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailAccountItem"
        "400":
          description: Invalid credentials or connection failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /email/status:
    get:
      operationId: getEmailStatus
      tags: [email]
      summary: Get connected email accounts
      description: Returns all connected email accounts for the current user
      responses:
        "200":
          description: Connected accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailStatusResponse"

  /email/disconnect:
    post:
      operationId: disconnectEmail
      tags: [email]
      summary: Disconnect an email account
      description: Disconnect a specific email account by sessionId, or all accounts if omitted
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisconnectEmailBody"
      responses:
        "200":
          description: Disconnected successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisconnectResponse"

  /email/scan:
    post:
      operationId: scanEmails
      tags: [email]
      summary: Scan emails for job applications
      description: Scan connected email inbox and extract job application data using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanEmailsBody"
      responses:
        "200":
          description: Scan completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanEmailsResponse"
        "400":
          description: No email connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all job applications
      description: Returns all tracked job applications, sorted by date descending
      parameters:
        - name: result
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a job application
      description: Manually create a new job application record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationBody"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /applications/summary:
    get:
      operationId: getApplicationsSummary
      tags: [applications]
      summary: Get application summary stats
      description: Returns counts by result status and recent activity
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationsSummary"

  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get a job application
      description: Returns a single job application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update a job application
      description: Update fields of an existing job application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationBody"
      responses:
        "200":
          description: Application updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete a job application
      description: Remove a job application record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ConnectEmailBody:
      type: object
      properties:
        provider:
          type: string
          description: "Email provider slug: gmail, outlook, yahoo, custom"
        email:
          type: string
        password:
          type: ["string", "null"]
          description: App password or IMAP password (required unless oauthToken is provided)
        oauthToken:
          type: ["string", "null"]
          description: Google OAuth access token (used instead of password for Gmail OAuth sign-in)
        imapHost:
          type: ["string", "null"]
          description: Required for custom provider
        imapPort:
          type: ["integer", "null"]
          description: Required for custom provider (e.g. 993)
      required:
        - provider
        - email

    EmailAccountItem:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        provider:
          type: string
        lastScanned:
          type: ["string", "null"]
          description: ISO 8601 timestamp of last scan
      required:
        - id
        - email
        - provider
        - lastScanned

    EmailStatusResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/EmailAccountItem"
      required:
        - accounts

    DisconnectEmailBody:
      type: object
      properties:
        sessionId:
          type: ["integer", "null"]
          description: ID of the specific session to disconnect. If omitted, all accounts are disconnected.

    DisconnectResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    ScanEmailsBody:
      type: object
      properties:
        maxEmails:
          type: ["integer", "null"]
          description: Maximum number of emails to scan (default 200)
        daysBack:
          type: ["integer", "null"]
          description: How many days back to look (default 90)
        clearPrevious:
          type: ["boolean", "null"]
          description: If true, delete applications outside the scan window before inserting new results (default true)
        oauthToken:
          type: ["string", "null"]
          description: Fresh Google OAuth access token (required when session uses OAuth auth)
      required: []

    ScanEmailsResponse:
      type: object
      properties:
        found:
          type: integer
          description: Number of job-related emails found
        added:
          type: integer
          description: Number of new applications added
        updated:
          type: integer
          description: Number of existing applications updated
      required:
        - found
        - added
        - updated

    Application:
      type: object
      properties:
        id:
          type: integer
        dateOfContact:
          type: string
          description: ISO 8601 date string
        position:
          type: ["string", "null"]
        employer:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
          description: Name of recruiter or hiring manager
        interviewerInfo:
          type: ["string", "null"]
          description: "For interview emails: interviewer name(s) and title(s)"
        methodOfContact:
          type: string
          description: "email, zoom, teams, google-meet, phone, linkedin, other"
        emailAddress:
          type: ["string", "null"]
        result:
          type: string
          description: "interview, next-stage, rejected, no-response"
        notes:
          type: ["string", "null"]
        sourceEmailId:
          type: ["string", "null"]
          description: Original email message ID for deduplication
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - dateOfContact
        - position
        - employer
        - contactName
        - methodOfContact
        - emailAddress
        - result
        - notes
        - sourceEmailId
        - createdAt
        - updatedAt

    CreateApplicationBody:
      type: object
      properties:
        dateOfContact:
          type: string
        position:
          type: ["string", "null"]
        employer:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        interviewerInfo:
          type: ["string", "null"]
        methodOfContact:
          type: string
        emailAddress:
          type: ["string", "null"]
        result:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - dateOfContact
        - methodOfContact
        - result

    UpdateApplicationBody:
      type: object
      properties:
        dateOfContact:
          type: string
        position:
          type: ["string", "null"]
        employer:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        interviewerInfo:
          type: ["string", "null"]
        methodOfContact:
          type: string
        emailAddress:
          type: ["string", "null"]
        result:
          type: string
        notes:
          type: ["string", "null"]

    ApplicationsSummary:
      type: object
      properties:
        total:
          type: integer
        byResult:
          type: object
          properties:
            interview:
              type: integer
            nextStage:
              type: integer
            rejected:
              type: integer
            noResponse:
              type: integer
          required:
            - interview
            - nextStage
            - rejected
            - noResponse
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/Application"
      required:
        - total
        - byResult
        - recentActivity